• fix(vpn-frame): recover standalone PN assignments after SN restart
  - use Unix-second PN versions across the command protocol
  - force a full sync after failed client reconciliation
  - add restart, retry, and protocol compatibility coverage

+++ vpn-frame/src/client/vpn_server_client.rs
@@ -89,9 +89,9 @@ impl<
     pub async fn get_vpn_info(
         &self,
         cur_version: Option<u16>,
-        cur_pn_info_version: Option<u16>,
+        cur_pn_info_version: Option<u32>,
         client_version: Option<String>,
-    ) -> VpnResult<((u16, u16), Vec<NodeVpnInfo>)> {
+    ) -> VpnResult<((u16, u32), Vec<NodeVpnInfo>)> {
         let req = GetVpnInfoReq {
             seq: self.gen_seq.generate(),
             info_version: cur_version,
